feat: add show trailer button
